requirements/dev.txt
-r common.txt
django-debug-toolbar==1.0.1
django_pdb==0.3.2
